FBP-48. Expanded undo to add/remove and properties

+++ xml-parser-api/src/main/kotlin/com/valb3r/bpmn/intellij/plugin/bpmn/api/events/EventInterfaces.kt
@@ -14,7 +14,7 @@ interface Event
 data class EventBlock(val size: Int)
 
 interface EventOrder<T: Event> {
-    val order: Long
+    val order: Int
     val event: T
     val block: EventBlock?
 }
